- removed cmfseo bundle - added sonata seo bundle - added seo on all controllers - toughened up security on all controllers - created adminVoterService - updated twig templates as well

+++ src/Controller/AssociationController.php
@@ -6,22 +6,39 @@ use App\Entity\Association;
 use App\Form\AssociationType;
 use App\Repository\AssociationRepository;
 use App\Repository\EventRepository;
+use App\Service\AdminVoterService;
 use App\Service\FileUploader;
+use Sonata\SeoBundle\Seo\SeoPageInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 /**
  * @Route("/association")
  */
 class AssociationController extends AbstractController
 {
+    private $seo;
+    public function __construct(SeoPageInterface $seo)
+    {
+        $this->seo = $seo;
+    }
+    
     /**
      * @Route("/", name="association_index", methods={"GET"})
      */
     public function index(AssociationRepository $associationRepository): Response
     {
+        $this->seo
+            ->addTitle('Associations & Clubs')
+            ->addMeta('name', 'description', 'These are the different clubs and associations found within Lae city.')
+            ->addMeta('property', 'og:title', 'Associations & Events')
+            ->addMeta('property', 'og:type', 'blog')
+            ->addMeta('property', 'og:url',  $this->generateUrl('association_index', [], UrlGeneratorInterface::ABSOLUTE_URL))
+            ->addMeta('property', 'og:description', 'These are the different clubs and associations found within Lae city.')
+        ;
         return $this->render('association/index.html.twig', [
             'associations' => $associationRepository->findAll(),
         ]);
@@ -30,8 +47,12 @@ class AssociationController extends AbstractController
     /**
      * @Route("/new", name="association_new", methods={"GET","POST"})
      */
-    public function new(Request $request, FileUploader $fileUploader): Response
+    public function new(Request $request, FileUploader $fileUploader, AdminVoterService $adminVoterService): Response
     {
+        if ($adminVoterService->isAdmin($this->getUser()->getRoles())) {
+            return $this->redirectToRoute('home');
+        }
+
         $association = new Association();
         $form = $this->createForm(AssociationType::class, $association);
         $form->handleRequest($request);
@@ -60,6 +81,14 @@ class AssociationController extends AbstractController
      */
     public function show(Association $association, EventRepository $eventRepository): Response
     {
+        $this->seo
+            ->addTitle($association->getName())
+            ->addMeta('name', 'description', $association->getDescription())
+            ->addMeta('property', 'og:title', $association->getName())
+            ->addMeta('property', 'og:type', 'blog')
+            ->addMeta('property', 'og:url',  $this->generateUrl('association_show', ['id'=>$association->getId()], UrlGeneratorInterface::ABSOLUTE_URL))
+            ->addMeta('property', 'og:description', $association->getDescription())
+        ;
         $events = $eventRepository->findBy(['publish'=>true, 'association'=>$association],['id'=>'desc']);
         return $this->render('association/show.html.twig', [
             'association' => $association,
@@ -70,8 +99,12 @@ class AssociationController extends AbstractController
     /**
      * @Route("/{id}/edit", name="association_edit", methods={"GET","POST"})
      */
-    public function edit(Request $request, Association $association): Response
+    public function edit(Request $request, Association $association, AdminVoterService $adminVoterService): Response
     {
+        if ($adminVoterService->isAdmin($this->getUser()->getRoles())) {
+            return $this->redirectToRoute('home');
+        }
+
         $form = $this->createForm(AssociationType::class, $association);
         $form->handleRequest($request);
 
@@ -90,8 +123,12 @@ class AssociationController extends AbstractController
     /**
      * @Route("/{id}", name="association_delete", methods={"DELETE"})
      */
-    public function delete(Request $request, Association $association): Response
+    public function delete(Request $request, Association $association, AdminVoterService $adminVoterService): Response
     {
+        if ($adminVoterService->isAdmin($this->getUser()->getRoles())) {
+            return $this->redirectToRoute('home');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$association->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($association);
